+++ core-commonui/unit-createreport/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/createreport/di/Utils.kt
@@ -1,0 +1,8 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.createreport.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.createreport.CreateReportMviModel
+
+expect fun getCreateReportViewModel(
+    postId: Int? = null,
+    commentId: Int? = null,
+): CreateReportMviModel
