+++ settings.gradle.kts
@@ -52,6 +52,7 @@ include(":core-commonui:unit-communityinfo")
 include(":core-commonui:unit-instanceinfo")
 include(":core-commonui:unit-remove")
 include(":core-commonui:unit-reportlist")
+include(":core-commonui:unit-saveditems")
+include(":core-commonui:unit-createreport")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
-include(":core-commonui:unit-saveditems")
